+++ src/header.js
@@ -1,7 +1,7 @@
 // ==UserScript==
 // @name       RPH Tools
 // @namespace  https://openuserjs.org/scripts/shuffyiosys/RPH_Tools
-// @version    4.4.2
+// @version    4.4.3
 // @description Adds extended settings to RPH
 // @match      https://chat.rphaven.com/
 // @copyright  (c)2014 shuffyiosys@github
@@ -10,6 +10,6 @@
 // @license    MIT
 // ==/UserScript==
 
-const VERSION_STRING = "4.4.2";
+const VERSION_STRING = "4.4.3";
 
 const SETTINGS_NAME = "rph_tools_settings";